# doc/openapi.yaml
openapi: 3.0.0
info:
  title: Music API
  version: 1.0.0
  description: REST API for managing music tracks
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /register:
    post:
      summary: Register an account
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
                email:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: Created a new account
        '400':
          description: Incorrect account info
        '500':
          description: Internal Server Error
  /login:
    post:
      summary: Log in to an existing account
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Incorrect Login credentials
        '500':
          description: Internal Server Error        
              
  /tracks:
    get:
      summary: List all tracks
      operationId: getAllTracks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sucessfully retrieved list
        '401':
          description: Unauthorized - Invalid Credentials
          
    post:
      summary: Create a new track
      operationId: createTrack
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
      responses:
        '201':
          description: Created
        '401':
          description: Unauthorized - Invalid Credentials
        '500':
          description: Internal Server Error
          
  /tracks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID of the track
        schema:
          type: integer
    get:
      summary: Get a track by ID
      operationId: getTrackById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A track object
        '401':
          description: Unauthorized - Invalid Credentials
        '404':
          description: Track not found
        '500':
          description: Internal Server Error
    put:
      summary: Update a track by ID
      operationId: updateTrack
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
      responses:
        '200':
          description: Updated
        '401':
          description: Unauthorized - Invalid Credentials
        '404':
          description: Track not found
        '500':
          description: Internal Server Error
    delete:
      summary: Delete a track by ID
      operationId: deleteBook
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No content
        '401':
          description: Unauthorized - Invalid Credentials
        '404':
          description: Track not found
        '500':
          description: Internal Server Error
components:
  schemas:
    Track:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        userID:
          type: string
          
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT